+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Mar 01 21:03:37 CET 2020
+gradle.version=5.2.1
