+++ src/main/java/Area.java
@@ -1,0 +1,3 @@
+public class Area {
+
+}
