+++ src/main/java/Player.java
@@ -1,0 +1,3 @@
+public class Player {
+
+}
