--- textAdventureGame/src/module-info.java
@@ -1,2 +1,0 @@
-module textAdventureGame {
-}
