--- textAdventureGame/src/textAdventureGame/Area.java
@@ -1,5 +1,0 @@
-package textAdventureGame;
-
-public class Area {
-
-}
