--- textAdventureGame/src/textAdventureGame/Player.java
@@ -1,5 +1,0 @@
-package textAdventureGame;
-
-public class Player {
-
-}
